refactor(office-preview): hoist base64ToArrayBuffer to shared
Move the byte-identical atob→Uint8Array→ArrayBuffer loop out of DocxViewer
and XlsxViewer into src/shared/zip-magic.ts (alongside the zip-magic helper
already hoisted there). One place to fix if the relay ever ships whitespace-
tolerant or URL-safe base64.
